Stop associating issues with modules directly instead link them by filepath
This ensures we don't slow down module-graph construction with a lot of root tasks and collectibles.  Instead we compute traces when we flatten issues into `PlainIssues` for display.

To do this we need to ensure that enough information is available when `get_plain_issues` is called, so i have emitted the full `SingleModuleGraph` as a collectible.

The import trace format is still a WIP but this plugs all the relevant pieces together